Fix build on x86_64
Patched by Kanru Chen <ckanru@gmail.com>


git-svn-id: https://eserv.googlecode.com/svn/trunk@6 a7618a5a-0e41-11df-9821-6b6fb8eff287

+++ libeserv/entry.h
@@ -30,7 +30,7 @@ extern "C" {
 #define	closesocket(a)		close(a)
 #define	ERRNO			errno
 #define	INVALID_SOCKET		(-1)
-typedef int SOCKET;
+typedef long SOCKET;
 
 #define  EX_SOCK_ERRNO errno
 #define	 EX_EINTR EINTR
